Transplant the comma_int functions into money.py as helper functions
